Default stripesize 64k & config file setting for it; Clear many compiler warnings (i386) & associated bugs - hopefully without introducing too many new bugs:-)  (Same exercise required for other archs.) Default compilation has optimisation - or else use ./configure --enable-debug

+++ lib/datastruct/bitset.c
@@ -12,8 +12,8 @@
 
 bitset_t bitset_create(struct pool *mem, unsigned num_bits)
 {
-	int n = (num_bits / BITS_PER_INT) + 2;
-	int size = sizeof(int) * n;
+	unsigned n = (num_bits / BITS_PER_INT) + 2;
+	size_t size = sizeof(int) * n;
 	unsigned *bs = pool_zalloc(mem, size);
 
 	if (!bs)
